requirements.txt
streamlit
python-dotenv
langchain
langchain-huggingface
huggingface_hub
transformers